Break down MCP route registrations
